datalake: moved some types to separate file
Moved basic type definitions to separate file to loose the dependencies
on `data_writer_interface.h`.
